Condições de erro consertadas e adição,remoçao.informação e lista de contatos funcionando

+++ backend/src/app.ts
@@ -5,7 +5,7 @@ import cors from 'cors';
 import logger from './logger';
 import setupRoutes from './routes/index';
 import { HttpError } from './utils/errors/http.error';
-import { FailureResult } from './utils/result';
+import { FailureResult, SuccessResult } from './utils/result';
 import Database from './database';
 import ContactsDatabase from './database/contacts.database';
 
@@ -39,15 +39,20 @@ app.get('/api/contacts', async (req, res, next) => {
 app.post('/api/contacts', async (req, res, next) => {
   try {
     const newContact = req.body; // O novo contato enviado no corpo da requisição
-    const contactsDatabase = ContactsDatabase.getInstance();
-    contactsDatabase.addContact(newContact); // Adiciona o novo contato ao banco de dados
+    
+    if (newContact && newContact.name && newContact.number) { // Verifica se o corpo da requisição contém os campos necessários
+      const contactsDatabase = ContactsDatabase.getInstance();
+      contactsDatabase.addContact(newContact); // Adiciona o novo contato ao banco de dados
 
-    res.status(201).json({ message: "Contato adicionado com sucesso" }); // Retorna 201 Created para indicar que o contato foi adicionado com sucesso
+      res.status(201).json({ message: "Contato adicionado com sucesso" }); // Retorna 201 Created para indicar que o contato foi adicionado com sucesso
+    } else {
+      res.status(400).json({ message: "Erro ao adicionar contato" }); // Se os campos necessários não estiverem presentes, retorna um erro 400 Bad Request
+    }
   } catch (error) {
+    res.status(500).json({ message: "Erro ao adicionar contato" });
     next(error); // Se houver um erro, passe para o próximo middleware de tratamento de erro
   }
 });
-
 
 // Rota para obter informações de um contato específico
 app.get('/api/contacts/:id', async (req, res, next) => {
@@ -71,12 +76,17 @@ app.delete('/api/contacts/delete/:id', async (req, res, next) => {
   try {
     const contactId = req.params.id;
     const contactsDatabase = ContactsDatabase.getInstance();
-    contactsDatabase.deleteContact(contactId); // Chama o método deleteContact passando o ID do contato a ser deletado
-    contactsDatabase.deleteContact(contactId);
 
-    res.status(204).json({msg:"Contato deletado com sucesso." }); // Retorna 204 No Content para indicar que o contato foi deletado com sucesso
+    // Verifica se o contato com o ID fornecido existe
+    const contactToDelete = contactsDatabase.getContactById(contactId);
+    if (contactToDelete) { 
+      contactsDatabase.deleteContact(contactId); // Se o contato existir, então deleta
+      res.status(201).json({ message: "Contato removido com sucesso" });
+
+    } else {
+      return res.status(404).json({ message: 'Contato não encontrado' }); // Se o contato não existir, retorna um erro 404
+    }
   } catch (error) {
-    res.status(500).json({msg: "Falha ao deletar o contato." });
     next(error);
   }
 });
